DRY up QBO searchers
